# docs/api.yaml
openapi: 3.0.3
info:
  title: LLM Knowledge Base API
  description: |
    本地运行的 LLM 知识库应用 API。

    支持导入多种来源（URL、PDF、视频、代码），自动编译成 Wiki 文章，并提供智能问答功能。
  version: 0.1.0
  contact:
    name: LLM Knowledge Base
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发服务器
...
paths:
...
  /documents:
    get:
      tags: [Documents]
      summary: 获取文档列表
      operationId: listDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [web, paper, video, code]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 文档列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
  /documents/{id}:
    get:
      tags: [Documents]
      summary: 获取文档详情
      operationId: getDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 文档详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: 文档不存在

    put:
      tags: [Documents]
      summary: 更新文档
      operationId: updateDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: 更新成功

    delete:
      tags: [Documents]
      summary: 删除文档
      operationId: deleteDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
  /documents/search:
    get:
      tags: [Documents]
      summary: 搜索文档
      operationId: searchDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'

  # ========== 编译接口 ==========
  /compile:
    post:
      tags: [Compile]
      summary: 触发编译
      description: |
        小批量（≤5篇）同步返回结果，大批量（>5篇）异步处理返回 task_id
      operationId: compileDocuments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompileRequest'
      responses:
        '200':
          description: 编译结果或任务ID
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CompileSyncResult'
                  - $ref: '#/components/schemas/CompileAsyncResult'
  /compile/tasks/{taskId}:
    get:
      tags: [Compile]
      summary: 查询编译任务状态
      operationId: getCompileTaskStatus
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 任务状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileTaskStatus'
  /compile/tasks:
    get:
      tags: [Compile]
      summary: 获取编译历史
      operationId: listCompileTasks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 编译任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileTaskList'

  # ========== 问答接口 ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [web, paper, video, code]
        status:
          type: string
          enum: [pending, processed]
        content:
          type: string
        source_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        related_docs:
          type: array
          items:
            type: string
    DocumentListResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
    DocumentSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
    UpdateDocumentRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, processed]
    SearchResult:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              snippet:
                type: string
              score:
                type: number
                format: float

    # ========== 编译相关 ==========
    CompileRequest:
      type: object
      required: [doc_ids]
      properties:
        doc_ids:
          type: array
          items:
            type: string
          minItems: 1
        options:
          type: object
          properties:
            generate_concepts:
              type: boolean
              default: true
            discover_links:
              type: boolean
              default: true
            output_language:
              type: string
              default: zh-CN
    CompileSyncResult:
      type: object
      properties:
        status:
          type: string
          enum: [completed]
        compiled_count:
          type: integer
        wiki_docs:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              path:
                type: string
    CompileAsyncResult:
      type: object
      properties:
        status:
          type: string
          enum: [queued]
        task_id:
          type: string
        total_docs:
          type: integer
    CompileTaskStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            failed:
              type: integer
            current_doc:
              type: string
        result:
          $ref: '#/components/schemas/CompileSyncResult'
    CompileTaskList:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              task_id:
                type: string
              status:
                type: string
              created_at:
                type: string
                format: date-time

    # ========== 问答相关 ==========
...